migrate lobby
