+++ src/pages/lobby/index.ts
@@ -1,0 +1,2 @@
+export * from "./init";
+export * from "./update";
