+++ src/pages/lobby/init.ts
@@ -1,0 +1,1 @@
+export function init(root: HTMLElement) {}
